Kafka fix

+++ order-service/src/main/java/com/example/orderservice/dto/OrderRequestDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class OrderRequestDTO {
     private String customerName;
     private String itemName;
-    private Double totalAmount;
+    private int totalAmount;
 }
